Conflicts:
	CHANGELOG.md

+++ src/umpire/interface/c_fortran/wrapResourceManager.h
@@ -90,12 +90,12 @@ size_t umpire_resourcemanager_get_size(umpire_resourcemanager * self,
 
 umpire_allocator * umpire_resourcemanager_make_allocator_pool(
     umpire_resourcemanager * self, const char * name,
-    umpire_allocator allocator, int initial_size, int block,
+    umpire_allocator allocator, size_t initial_size, size_t block,
     umpire_allocator * SHC_rv);
 
 umpire_allocator * umpire_resourcemanager_make_allocator_bufferify_pool(
     umpire_resourcemanager * self, const char * name, int Lname,
-    umpire_allocator allocator, int initial_size, int block,
+    umpire_allocator allocator, size_t initial_size, size_t block,
     umpire_allocator * SHC_rv);
 
 #ifdef __cplusplus
